% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Damage and distance done by fires to dams}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{DISTANCE}{distance covered}
\item{DAMAGE}{cost of damage}D
...
}
}
\source{
Mastering Statistics, 8th Edition
}
\usage{
fire
}
\description{
A dataset containing the damage and distance covered to dams by fires.
}
\keyword{datasets}
