% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753talk0004.R
\docType{package}
\name{MATH4753talk0004}
\alias{MATH4753talk0004}
\title{Lab Function Package}
\description{
Contains functions used for MATH4753.
}
\author{
Grant Talkington \email{gtalkington@ou.edu}
}
