% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Generate a plot of LENGTH vs WEIGHT of the DDT dataset.}
\usage{
myddt(df, spec)
}
\arguments{
\item{df}{the dataframe to be read in (ddt.csv)}

\item{spec}{species of fish to be plotted}
}
\value{
a ggplot of the given species

a named list of DDT data frame before and after subsetting and rel freq table of RIVER before subsetting
}
\description{
Generate a plot of LENGTH vs WEIGHT of the DDT dataset.
}
\examples{
\dontrun{
myddt(ddt, "CCATFISH")
myddt(ddt, "SMBUFFALO")
myddt(ddt, "LMBASS")
}
}
