% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Create a density curve and area polygon of a given normal distribution.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Average of the distribution}

\item{sigma}{Standard deviation of the distribution}

\item{a}{Upper bound of the probability}
}
\value{
A density curve of the given distribution and probability
}
\description{
Create a density curve and area polygon of a given normal distribution.
}
\examples{
\dontrun{
myncurve(10, 3, 7)
myncurve(-2, 1.5, -3)
}
}
