% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbin.R
\name{simbin}
\alias{simbin}
\title{Create a barplot showing a binomial simulation iterated a given number of times}
\usage{
simbin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iteration samples}

\item{n}{size of sample}

\item{p}{probability of success}
}
\value{
A barplot showing the generated binomial simulation
}
\description{
Create a barplot showing a binomial simulation iterated a given number of times
}
\examples{
\dontrun{
simbin(20, 20, 0.8)
simbin(1000, 16, 0.3)
}
}
